code

+++ Arduinocode/receiver/receiver.h
@@ -20,7 +20,7 @@ typedef struct but_struct{
 typedef struct trans_data_struct
 {
   send_rec_e trans;      //type of data, command
-  int hndshk_sig;        //handshake value
+  byte hndshk_sig;        //handshake value
   but_s but;
 } trans_data_s ;//= {HNDSHK, 0,{COMMAND,NOUPDATE,LOW}};
 #endif
